Move TOKEN_CATEGORIES to shared-pure to fix token-ui build
token-ui's vite build failed because AbstractTokenForm imported
TOKEN_CATEGORIES at runtime from @l2beat/token-backend, whose CJS-only
build is not processed by vite's commonjs interop for workspace
packages. Declare the constant in @l2beat/shared-pure (dual ESM/CJS)
and import it from there instead.
