+++ packages/database/src/repositories/TokenMetadataRepository.ts
@@ -1,17 +1,9 @@
+import type { TokenCategory } from '@l2beat/shared-pure'
 import type { Insertable, Selectable } from 'kysely'
 import { BaseRepository } from '../BaseRepository'
 import type { TokenMetadata } from '../kysely/generated/types'
 
-export const TOKEN_CATEGORIES = [
-  'ether',
-  'stablecoin',
-  'btc',
-  'rwaRestricted',
-  'rwaPublic',
-  'other',
-] as const
-
-export type TokenCategory = (typeof TOKEN_CATEGORIES)[number]
+export type { TokenCategory }
 
 export type TokenSource =
   | 'canonical'
