+++ packages/shared-pure/src/types/index.ts
@@ -17,6 +17,7 @@ export * from './ProjectId.js'
 export * from './ProjectValueType.js'
 export * from './StringWithAutocomplete.js'
 export * from './Token.js'
+export * from './TokenCategory.js'
 export * from './TokenId.js'
 export * from './TrackedTxsConfigSubtype.js'
 export * from './TrackedTxsConfigType.js'
