+++ packages/token-backend/src/index.ts
@@ -1,4 +1,3 @@
-export { TOKEN_CATEGORIES, type TokenCategory } from '@l2beat/database'
 export { getTokenDbClient, type TokenDbClient } from './client'
 export type { Command } from './commands'
 export type {
